Added DependencyResolver.get_channel_solution() method that permits to calculate the solution for a given action.
A new Enum, ActionType has been added to easily specify the action (ENABLE/DISABLE).

In order to make the DISABLE action working, get_channel_blockers() has been modified to take the
action to calculate the blockers for.
If the action is DISABLE, blockers will be calculated as a Conflict on the package the channel to remove is marked as a dependency.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>